Changes are logged inside changes_log.txt: soft modification of some files in order to compile, build, and run successfully the whole codebase

+++ template/template_types.h
@@ -1,3 +1,3 @@
-typedef long EVALUATION_TYPE;
+typedef long long EVALUATION_TYPE;
 typedef int GENOME_TYPE;
 typedef int FITNESS_TYPE;
